[AN] refactor(#424): Token 관련 패키지 변경
