+++ android/gitudy/data/src/main/java/com/takseha/data/token/NavigationHandler.kt
@@ -1,4 +1,4 @@
-package com.takseha.data
+package com.takseha.data.token
 
 interface NavigationHandler {
     fun navigateToLogin()
